+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # nvencFFX - Changelog
 
+## [1.5.5] - 2025-10-25
+- Added "FPS Mode";
+- Added "Cancel" for "Play 10s Preview" button;
+- Added "Deshake" filter button.
+
 ## [1.5.4] - 2025-10-18
 - Screen recording now respects the "Preset" parameter value;
 - The clipboard is now preserved even after closing the program (for example, a copied ffmpeg command).